+++ backends/backends.go
@@ -7,11 +7,12 @@ import (
 )
 
 type Backend interface {
-	WriteMetadata(ctx context.Context, hash string, data map[string]string) (map[string]string, error)
+	WriteMetadata(ctx context.Context, hash string, key string, value io.ReadSeeker) error
 	ReadMetadata(ctx context.Context, hash string, keys []string) (map[string]string, error)
 
 	StoreArtifacts(ctx context.Context, storage localstorage.ReadableStorage, hash string, paths []string) ([]string, error)
-	FetchArtifacts(ctx context.Context, hash string, writeFile WriteFile) error
+	FetchArtifacts(ctx context.Context, hash string, readFile ReadFile, writeFile WriteFile) error
 }
 
+type ReadFile func(ctx context.Context, relPath string) (io.ReadSeekCloser, error)
 type WriteFile func(ctx context.Context, relPath string, content io.Reader) error