+++ backends/s3/backend.go
@@ -5,12 +5,14 @@ import (
 	"cas/localstorage"
 	"cas/tracing"
 	"context"
+	"crypto/sha1"
 	"errors"
 	"fmt"
+	"io"
 	"io/ioutil"
+	"os"
 	"path"
 	"path/filepath"
-	"strings"
 	"sync"
 	"time"
 
@@ -70,61 +72,25 @@ func EnsureBucket(ctx context.Context, cfg S3Config) error {
 	return err
 }
 
-func (s *S3Backend) WriteMetadata(ctx context.Context, hash string, data map[string]string) (map[string]string, error) {
+func (s *S3Backend) WriteMetadata(ctx context.Context, hash string, key string, value io.ReadSeeker) error {
 	ctx, span := tr.Start(ctx, "write_metadata")
 	defer span.End()
 
-	found, err := s.hasMetadata(ctx, hash, backends.MetadataTimeStamp)
-	if err != nil {
-		return nil, tracing.Error(span, err)
-	}
-	if !found {
-		data[backends.MetadataTimeStamp] = fmt.Sprintf("%v", time.Now().Unix())
-	}
+	span.SetAttributes(
+		attribute.String("key", key),
+	)
 
-	wg := sync.WaitGroup{}
-	wg.Add(len(data))
-
-	errChan := make(chan error, len(data))
-	writtenChan := make(chan pair, len(data))
-
-	for k, v := range data {
-
-		go func(c context.Context, key, value string) {
-			ctx, span := tr.Start(c, "write_"+key)
-			defer span.End()
-			defer wg.Done()
-
-			span.SetAttributes(
-				attribute.String("key", key),
-				attribute.String("value", value),
-			)
-
-			req := &s3.PutObjectInput{
-				Bucket: &s.cfg.BucketName,
-				Key:    s.metadataPath(hash, key),
-				Body:   strings.NewReader(value),
-			}
-
-			if _, err := s.client.PutObject(ctx, req); err != nil {
-				errChan <- tracing.Error(span, err)
-				return
-			}
-
-			writtenChan <- pair{key, value}
-
-		}(ctx, k, v)
+	req := &s3.PutObjectInput{
+		Bucket: &s.cfg.BucketName,
+		Key:    s.metadataPath(hash, key),
+		Body:   value,
 	}
 
-	wg.Wait()
-
-	written := collectMap(writtenChan)
-
-	if err := collectErrors(errChan); err != nil {
-		return written, tracing.Error(span, err)
+	if _, err := s.client.PutObject(ctx, req); err != nil {
+		return tracing.Error(span, err)
 	}
 
-	return written, nil
+	return nil
 }
 
 func (s *S3Backend) ReadMetadata(ctx context.Context, hash string, keys []string) (map[string]string, error) {
@@ -240,9 +206,19 @@ func (s *S3Backend) StoreArtifacts(ctx context.Context, storage localstorage.Rea
 	ctx, span := tr.Start(ctx, "store_artifacts")
 	defer span.End()
 
-	// force the timestamp to exist
-	if _, err := s.WriteMetadata(ctx, hash, map[string]string{}); err != nil {
+	_, found, err := backends.ReadTimestamp(ctx, s, hash)
+	if err != nil {
 		return nil, tracing.Error(span, err)
+	}
+
+	span.SetAttributes(attribute.Bool("has_timestamp", found))
+
+	if !found {
+		if err := backends.CreateHash(ctx, s, hash, time.Now()); err != nil {
+			return nil, tracing.Error(span, err)
+		}
+
+		span.SetAttributes(attribute.Bool("hash_created", true))
 	}
 
 	wg := sync.WaitGroup{}
@@ -271,10 +247,26 @@ func (s *S3Backend) StoreArtifacts(ctx context.Context, storage localstorage.Rea
 			}
 			defer content.Close()
 
+			sha, err := hashFile(ctx, content)
+			if err != nil {
+				errChan <- tracing.Error(span, err)
+				return
+			}
+
+			span.SetAttributes(attribute.String("local_hash", sha))
+
+			if _, err := content.Seek(0, 0); err != nil {
+				errChan <- tracing.Error(span, err)
+				return
+			}
+
 			req := &s3.PutObjectInput{
 				Bucket: &s.cfg.BucketName,
 				Key:    &s3path,
 				Body:   content,
+				Metadata: map[string]string{
+					"sha1": sha,
+				},
 			}
 
 			if _, err := s.client.PutObject(ctx, req); err != nil {
@@ -297,7 +289,17 @@ func (s *S3Backend) StoreArtifacts(ctx context.Context, storage localstorage.Rea
 	return written, nil
 }
 
-func (s *S3Backend) FetchArtifacts(ctx context.Context, hash string, writeFile backends.WriteFile) error {
+func hashFile(ctx context.Context, file io.Reader) (string, error) {
+
+	hash := sha1.New()
+	if _, err := io.Copy(hash, file); err != nil {
+		return "", err
+	}
+
+	return fmt.Sprintf("%x", hash.Sum(nil)), nil
+}
+
+func (s *S3Backend) FetchArtifacts(ctx context.Context, hash string, readFile backends.ReadFile, writeFile backends.WriteFile) error {
 	ctx, span := tr.Start(ctx, "fetch_artifacts")
 	defer span.End()
 
@@ -312,28 +314,73 @@ func (s *S3Backend) FetchArtifacts(ctx context.Context, hash string, writeFile b
 	errChan := make(chan error, len(paths))
 
 	for _, p := range paths {
-		go func(ctx context.Context, filePath string) {
-			ctx, span := tr.Start(ctx, "fetch_"+path.Base(filePath))
+		go func(ctx context.Context, localPath string) {
+			ctx, span := tr.Start(ctx, "fetch_"+path.Base(localPath))
 			defer span.End()
 			defer wg.Done()
 
-			remotePath := s.artifactPath(hash, filePath)
-			span.SetAttributes(attribute.String("remote_path", remotePath))
+			remotePath := s.artifactPath(hash, localPath)
+			span.SetAttributes(
+				attribute.String("local_path", localPath),
+				attribute.String("remote_path", remotePath),
+			)
 
-			res, err := s.client.GetObject(ctx, &s3.GetObjectInput{
-				Bucket: &s.cfg.BucketName,
-				Key:    &remotePath,
-			})
-			if err != nil {
+			localContent, err := readFile(ctx, localPath)
+			if err != nil && !errors.Is(err, os.ErrNotExist) {
 				errChan <- tracing.Error(span, err)
 				return
 			}
 
-			defer res.Body.Close()
+			span.SetAttributes(attribute.Bool("has_local_version", localContent != nil))
+			downloadFile := true
 
-			if err := writeFile(ctx, filePath, res.Body); err != nil {
-				errChan <- tracing.Error(span, err)
-				return
+			if localContent != nil {
+				defer localContent.Close()
+				localHash, err := hashFile(ctx, localContent)
+				if err != nil {
+					errChan <- tracing.Error(span, err)
+					return
+				}
+
+				span.SetAttributes(attribute.String("local_hash", localHash))
+
+				r, err := s.client.HeadObject(ctx, &s3.HeadObjectInput{
+					Bucket: &s.cfg.BucketName,
+					Key:    &remotePath,
+				})
+				if err != nil {
+					errChan <- tracing.Error(span, err)
+					return
+				}
+
+				remoteHash, hasRemoteHash := r.Metadata["sha1"]
+				span.SetAttributes(
+					attribute.Bool("has_remote_hash", hasRemoteHash),
+					attribute.String("remote_hash", remoteHash),
+				)
+
+				downloadFile = !hasRemoteHash || remoteHash != localHash
+
+			}
+
+			span.SetAttributes(attribute.Bool("download_file", downloadFile))
+
+			if downloadFile {
+				res, err := s.client.GetObject(ctx, &s3.GetObjectInput{
+					Bucket: &s.cfg.BucketName,
+					Key:    &remotePath,
+				})
+				if err != nil {
+					errChan <- tracing.Error(span, err)
+					return
+				}
+
+				defer res.Body.Close()
+
+				if err := writeFile(ctx, localPath, res.Body); err != nil {
+					errChan <- tracing.Error(span, err)
+					return
+				}
 			}
 
 		}(ctx, p)