+++ command/artifact.go
@@ -38,6 +38,10 @@ func (c *ArtifactCommand) Flags() *pflag.FlagSet {
 	return flags
 }
 
+func (c *ArtifactCommand) EnvironmentVariables() map[string]string {
+	return map[string]string{}
+}
+
 func (c *ArtifactCommand) RunContext(ctx context.Context, args []string) error {
 
 	ctx, span := c.tr.Start(ctx, "run")